requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
pillow==10.2.0
openai>=1.51.0
requests==2.31.0
python-multipart==0.0.6
jinja2==3.1.3
python-dotenv==1.0.1
pydantic==2.6.1
psycopg2-binary==2.9.9
gunicorn==21.2.0
PyPDF2==3.0.1
websockets==12.0
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==4.0.1
redis==5.0.1
openpyxl==3.1.5
xlrd==1.2.0
xlwt==1.3.0
xlutils==2.0.0
